Add binder (#14)
* feat(types): add binder types

* fix(parser): consume equals token in assignment

* feat(binder): add binder

* feat(types): add new diagnostic codes and source

Some new diagnostic codes have been added for errors which can be
reported by the binder. A new diagnostic source has also been added
specifically for the binder.

BREAKING CHANGE: The diagnostic source value for `Checker` has changed
from `1 << 2` to `1 << 3`.

* feat(binder): add diagnostics and improve binding

* Diagnostic messages are now generated by the binder in situations
where it finds duplicate or missing symbols.

* References are no longer added to the original symbol when a duplicate
of it is found.

* Some nodes are assigned the symbol of one of their children without
another reference being added to that symbol. This is simply for
convenience later on so visitors don't have to go as deep to get symbol
information.