feat(syntax): add identifier and number nodes
This change refactors the codebase to fully separate tokens from other
syntax nodes. Previously, the `IdentifierLiteral` and `NumberLiteral`
node types were being used to represent both a number/identifier token
as well as a number/token node which could appear as a child of another
node. This lead to some ugly code in the parser and the type
definitions.

The code has been refactored so that identifiers and numbers now have
their own node types instead of using the token types. Test cases have
been updated where necessary to reflect this refactoring.