feat(semantic): fix typechecking and add data flow
This commit refactors the typechecker so each node's typechecking logic
lives in the same file as the node definition.

The parser has also been refactored to build a control flow graph of the
program as it is being parsed. This process also involved:

+ The removal of the StopStatement in favour of a GotoStatement.
+ The addition of a "block threading" optimiser. The parser generates a
lot of unnecessary blocks while building the graph, so it is the block
threader's job to throw those unnecessary blocks away without
functionality affecting the code.
+ The addition of a "linked symbol table" to the block statement node.
This symbol table is used to retain scoping information about the blocks
after they have been converted into a CFG.

Some more information has been added to the types to represent the
operators that are supported by that type.

A new pass which runs after the type checking called the `DataFlowPass`
checks for things such as unused variables, unreachable code, infinite
loops, etc.